合同ロジックを作成

+++ src/geometry/triangle/sides_ratio.rs
@@ -1,13 +1,13 @@
 use super::sides;
 
 pub struct SidesRatio {
-    a: f32,
-    b: f32,
-    c: f32,
+    a: f64,
+    b: f64,
+    c: f64,
 }
 
 impl SidesRatio {
-    pub fn new(a: f32, b: f32, c: f32) -> SidesRatio {
+    pub fn new(a: f64, b: f64, c: f64) -> SidesRatio {
         SidesRatio { a, b, c }
     }
 
@@ -15,13 +15,13 @@ impl SidesRatio {
         sides::is_valid(self.a, self.b, self.c)
     }
 
-    pub fn a(&self) -> f32 {
+    pub fn a(&self) -> f64 {
         self.a
     }
-    pub fn b(&self) -> f32 {
+    pub fn b(&self) -> f64 {
         self.b
     }
-    pub fn c(&self) -> f32 {
+    pub fn c(&self) -> f64 {
         self.c
     }
 }
